Web: Bump core dependencies and use dpdm instead of madge
