+++ mwdb/web/package.json
@@ -3,12 +3,12 @@
   "version": "2.8.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.0-14",
-    "@fortawesome/free-regular-svg-icons": "^5.1.0-11",
-    "@fortawesome/free-solid-svg-icons": "^5.1.0-11",
-    "@fortawesome/react-fontawesome": "0.1.0-11",
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-regular-svg-icons": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@fortawesome/react-fontawesome": "0.1.18",
     "@mwdb-web/commons": "file:src/commons",
-    "axios": "^0.21.1",
+    "axios": "^0.27.2",
     "bootstrap": "4.3.1",
     "bootstrap-select": "^1.13.18",
     "d3": "^5.10.0",
@@ -19,37 +19,37 @@
     "jquery": "^3.3.1",
     "lodash": "^4.17.10",
     "popper.js": "^1.16.1",
-    "query-string": "^5.1.1",
-    "react": "^16.13.1",
-    "react-ace": "^9.4.3",
+    "query-string": "^7.1.1",
+    "react": "^18.1.0",
+    "react-ace": "^10.1.0",
     "react-copy-to-clipboard": "^5.0.2",
-    "react-dom": "^16.4.1",
-    "react-dropzone": "^11.3.1",
-    "react-google-recaptcha": "^2.0.1",
-    "react-infinite-scroller": "^1.2.0",
-    "react-js-pagination": "^3.0.2",
-    "react-modal": "^3.5.1",
+    "react-dom": "^18.1.0",
+    "react-dropzone": "^14.2.1",
+    "react-google-recaptcha": "^2.1.0",
+    "react-infinite-scroller": "^1.2.6",
+    "react-js-pagination": "^3.0.3",
+    "react-modal": "^3.15.1",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "readable-timestamp": "^0.2.0",
     "sha1": "^1.1.1",
-    "swagger-ui-react": "3.52.5"
+    "swagger-ui-react": "^4.11.1"
   },
   "scripts": {
     "start": "react-app-rewired start",
-    "build": "madge --circular --extensions js src/ && react-app-rewired build",
+    "build": "dpdm --exit-code circular:1 --tree false --progress false --warning false ./src/index.js && react-app-rewired build",
     "test": "react-app-rewired test --env=jsdom",
     "eject": "react-scripts eject"
   },
   "devDependencies": {
     "customize-cra": "^1.0.0",
+    "dpdm": "^3.9.0",
     "exports-loader": "^1.1.0",
     "http-proxy-middleware": "^0.20.0",
     "imports-loader": "^1.1.0",
-    "madge": "^4.0.2",
     "prettier": "^2.4.1",
     "react-app-rewired": "2.1.3",
-    "react-scripts": "4.0.3"
+    "react-scripts": "5.0.1"
   },
   "browserslist": {
     "production": [
